# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Initiative Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type InitiativeStatus = typeof InitiativeStatus[keyof typeof InitiativeStatus];


export const InitiativeStatus = {
  on_track: 'on_track',
  at_risk: 'at_risk',
  delayed: 'delayed',
  completed: 'completed',
  not_started: 'not_started',
} as const;

export type InitiativePriority = typeof InitiativePriority[keyof typeof InitiativePriority];


export const InitiativePriority = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface Initiative {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  status: InitiativeStatus;
  /**
     * @minimum 0
     * @maximum 100
     */
  progress: number;
  priority: InitiativePriority;
  /** @nullable */
  owner?: string | null;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  endDate?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type InitiativeInputStatus = typeof InitiativeInputStatus[keyof typeof InitiativeInputStatus];


export const InitiativeInputStatus = {
  on_track: 'on_track',
  at_risk: 'at_risk',
  delayed: 'delayed',
  completed: 'completed',
  not_started: 'not_started',
} as const;

export type InitiativeInputPriority = typeof InitiativeInputPriority[keyof typeof InitiativeInputPriority];


export const InitiativeInputPriority = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface InitiativeInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  status: InitiativeInputStatus;
  /**
     * @minimum 0
     * @maximum 100
     */
  progress: number;
  priority: InitiativeInputPriority;
  owner?: string;
  department?: string;
  startDate?: string;
  endDate?: string;
}

export type InitiativeUpdateStatus = typeof InitiativeUpdateStatus[keyof typeof InitiativeUpdateStatus];


export const InitiativeUpdateStatus = {
  on_track: 'on_track',
  at_risk: 'at_risk',
  delayed: 'delayed',
  completed: 'completed',
  not_started: 'not_started',
} as const;

export type InitiativeUpdatePriority = typeof InitiativeUpdatePriority[keyof typeof InitiativeUpdatePriority];


export const InitiativeUpdatePriority = {
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface InitiativeUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  status?: InitiativeUpdateStatus;
  /**
     * @minimum 0
     * @maximum 100
     */
  progress?: number;
  priority?: InitiativeUpdatePriority;
  owner?: string;
  department?: string;
  startDate?: string;
  endDate?: string;
}

export interface InitiativeProgressUpdate {
  id: number;
  initiativeId: number;
  note: string;
  author: string;
  createdAt: string;
}

export interface InitiativeProgressUpdateInput {
  /** @minLength 1 */
  note: string;
  /** @minLength 1 */
  author: string;
}

export interface DepartmentCount {
  department: string;
  count: number;
}

export interface PriorityCount {
  priority: string;
  count: number;
}

export interface DashboardSummary {
  total: number;
  onTrack: number;
  atRisk: number;
  delayed: number;
  completed: number;
  notStarted: number;
  avgProgress: number;
  byDepartment: DepartmentCount[];
  byPriority: PriorityCount[];
}

export interface RecentActivityItem {
  id: number;
  initiativeId: number;
  initiativeTitle: string;
  note: string;
  author: string;
  createdAt: string;
}

export interface Department {
  id: number;
  name: string;
}

export interface DepartmentInput {
  /** @minLength 1 */
  name: string;
}

export type ListInitiativesParams = {
status?: string;
department?: string;
priority?: string;
};

export type GetRecentActivityParams = {
limit?: number;
};