# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Initiative Tracker API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: initiatives
    description: Initiative CRUD
  - name: updates
    description: Initiative progress updates
  - name: dashboard
    description: Aggregated dashboard data
  - name: departments
    description: Department management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /initiatives:
    get:
      operationId: listInitiatives
      tags: [initiatives]
      summary: List all initiatives
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: department
          in: query
          required: false
          schema:
            type: string
        - name: priority
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of initiatives
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Initiative"
    post:
      operationId: createInitiative
      tags: [initiatives]
      summary: Create a new initiative
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitiativeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Initiative"

  /initiatives/{id}:
    get:
      operationId: getInitiative
      tags: [initiatives]
      summary: Get a single initiative
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Initiative found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Initiative"
        "404":
          description: Not found
    patch:
      operationId: updateInitiative
      tags: [initiatives]
      summary: Update an initiative
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitiativeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Initiative"
        "404":
          description: Not found
    delete:
      operationId: deleteInitiative
      tags: [initiatives]
      summary: Delete an initiative
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  /initiatives/{id}/updates:
    get:
      operationId: listInitiativeUpdates
      tags: [updates]
      summary: List progress updates for an initiative
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of updates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InitiativeProgressUpdate"
    post:
      operationId: createInitiativeUpdate
      tags: [updates]
      summary: Add a progress update to an initiative
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitiativeProgressUpdateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InitiativeProgressUpdate"

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get aggregated dashboard statistics
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /dashboard/recent-activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Get recent updates across all initiatives
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Recent activity feed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RecentActivityItem"

  /departments:
    get:
      operationId: listDepartments
      tags: [departments]
      summary: List all departments
      responses:
        "200":
          description: List of departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: createDepartment
      tags: [departments]
      summary: Create a department
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Initiative:
      type: object
      required: [id, title, status, progress, priority, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [on_track, at_risk, delayed, completed, not_started]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        priority:
          type: string
          enum: [high, medium, low]
        owner:
          type: ["string", "null"]
        department:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string

    InitiativeInput:
      type: object
      required: [title, status, progress, priority]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        status:
          type: string
          enum: [on_track, at_risk, delayed, completed, not_started]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        priority:
          type: string
          enum: [high, medium, low]
        owner:
          type: string
        department:
          type: string
        startDate:
          type: string
        endDate:
          type: string

    InitiativeUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        status:
          type: string
          enum: [on_track, at_risk, delayed, completed, not_started]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        priority:
          type: string
          enum: [high, medium, low]
        owner:
          type: string
        department:
          type: string
        startDate:
          type: string
        endDate:
          type: string

    InitiativeProgressUpdate:
      type: object
      required: [id, initiativeId, note, author, createdAt]
      properties:
        id:
          type: integer
        initiativeId:
          type: integer
        note:
          type: string
        author:
          type: string
        createdAt:
          type: string

    InitiativeProgressUpdateInput:
      type: object
      required: [note, author]
      properties:
        note:
          type: string
          minLength: 1
        author:
          type: string
          minLength: 1

    DashboardSummary:
      type: object
      required: [total, onTrack, atRisk, delayed, completed, notStarted, avgProgress, byDepartment, byPriority]
      properties:
        total:
          type: integer
        onTrack:
          type: integer
        atRisk:
          type: integer
        delayed:
          type: integer
        completed:
          type: integer
        notStarted:
          type: integer
        avgProgress:
          type: number
        byDepartment:
          type: array
          items:
            $ref: "#/components/schemas/DepartmentCount"
        byPriority:
          type: array
          items:
            $ref: "#/components/schemas/PriorityCount"

    DepartmentCount:
      type: object
      required: [department, count]
      properties:
        department:
          type: string
        count:
          type: integer

    PriorityCount:
      type: object
      required: [priority, count]
      properties:
        priority:
          type: string
        count:
          type: integer

    RecentActivityItem:
      type: object
      required: [id, initiativeId, initiativeTitle, note, author, createdAt]
      properties:
        id:
          type: integer
        initiativeId:
          type: integer
        initiativeTitle:
          type: string
        note:
          type: string
        author:
          type: string
        createdAt:
          type: string

    Department:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string

    DepartmentInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1