# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Initiative Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { DepartmentCount } from './departmentCount';
import type { PriorityCount } from './priorityCount';

export interface DashboardSummary {
  total: number;
  onTrack: number;
  atRisk: number;
  delayed: number;
  completed: number;
  notStarted: number;
  avgProgress: number;
  byDepartment: DepartmentCount[];
  byPriority: PriorityCount[];
}