# lib/api-zod/src/generated/types/initiative.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Initiative Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { InitiativePriority } from './initiativePriority';
import type { InitiativeStatus } from './initiativeStatus';

export interface Initiative {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  status: InitiativeStatus;
  /**
     * @minimum 0
     * @maximum 100
     */
  progress: number;
  priority: InitiativePriority;
  /** @nullable */
  owner?: string | null;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  endDate?: string | null;
  createdAt: string;
  updatedAt: string;
}