# lib/api-zod/src/generated/types/initiativeInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Initiative Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { InitiativeInputPriority } from './initiativeInputPriority';
import type { InitiativeInputStatus } from './initiativeInputStatus';

export interface InitiativeInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  status: InitiativeInputStatus;
  /**
     * @minimum 0
     * @maximum 100
     */
  progress: number;
  priority: InitiativeInputPriority;
  owner?: string;
  department?: string;
  startDate?: string;
  endDate?: string;
}