# lib/api-zod/src/generated/types/initiativeProgressUpdateInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Initiative Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface InitiativeProgressUpdateInput {
  /** @minLength 1 */
  note: string;
  /** @minLength 1 */
  author: string;
}