# lib/api-zod/src/generated/types/recentActivityItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Initiative Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface RecentActivityItem {
  id: number;
  initiativeId: number;
  initiativeTitle: string;
  note: string;
  author: string;
  createdAt: string;
}